split runtime input resources e2e test
